requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
easyocr==1.7.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
instructor==1.10.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
openai==1.97.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
